requirements.txt
# requirements.txt

streamlit==1.46.0
requests==2.32.3
pandas==2.2.3
mysql-connector-python==9.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
pydantic==2.11.5
pydantic_core==2.33.2
uvicorn==0.34.2
pytest==8.4.0
python-dotenv==1.0.1